requirements.txt
async-generator==1.10
attrs==22.1.0
brotlipy==0.7.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.4
cryptography==3.4.8
exceptiongroup==1.0.4
fake-useragent==0.1.14
h11==0.14.0
idna==3.4
outcome==1.2.0
pycparser==2.21
pyOpenSSL==20.0.1
PySocks==1.7.1
python-gnupg==0.5.0
requests==2.28.1
selenium==4.6.0
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.13
wheel==0.37.1
wsproto==1.2.0
